defect fixes

+++ src/main/java/com/agg/application/entity/ClaimLabor.java
@@ -33,7 +33,7 @@ public class ClaimLabor {
 	private String laborDescr;
 	
 	@Column(name="labor_hrs")
-	private int laborHrs;
+	private double laborHrs;
 	
 	private double rate;
 
@@ -100,18 +100,19 @@ public class ClaimLabor {
 	public void setLaborDescr(String laborDescr) {
 		this.laborDescr = laborDescr;
 	}
-
+	
+	
 	/**
 	 * @return the laborHrs
 	 */
-	public int getLaborHrs() {
+	public double getLaborHrs() {
 		return laborHrs;
 	}
 
 	/**
 	 * @param laborHrs the laborHrs to set
 	 */
-	public void setLaborHrs(int laborHrs) {
+	public void setLaborHrs(double laborHrs) {
 		this.laborHrs = laborHrs;
 	}
