+++ src/main/java/com/agg/application/model/ClaimPartDO.java
@@ -12,7 +12,7 @@ public class ClaimPartDO {
 	private int claimId;
 	private String partNo;
 	private String partDescr;
-	private int qty;
+	private double qty;
 	private double unitPrice;
 	private String partTotal;
 	/**
@@ -70,16 +70,17 @@ public class ClaimPartDO {
 	public void setPartDescr(String partDescr) {
 		this.partDescr = partDescr;
 	}
+	
 	/**
 	 * @return the qty
 	 */
-	public int getQty() {
+	public double getQty() {
 		return qty;
 	}
 	/**
 	 * @param qty the qty to set
 	 */
-	public void setQty(int qty) {
+	public void setQty(double qty) {
 		this.qty = qty;
 	}
 	/**
