+++ src/main/java/com/agg/application/vo/ClaimLabourVO.java
@@ -4,7 +4,7 @@ public class ClaimLabourVO {
 	private int id;
 	private String laborNo;
 	private String laborDescr;
-	private int laborHrs;
+	private double laborHrs;
 	private double laborHourlyRate;
 	/**
 	 * 
@@ -52,16 +52,18 @@ public class ClaimLabourVO {
 	public void setLaborDescr(String laborDescr) {
 		this.laborDescr = laborDescr;
 	}
+	
 	/**
 	 * @return the laborHrs
 	 */
-	public int getLaborHrs() {
+	public double getLaborHrs() {
 		return laborHrs;
 	}
+
 	/**
 	 * @param laborHrs the laborHrs to set
 	 */
-	public void setLaborHrs(int laborHrs) {
+	public void setLaborHrs(double laborHrs) {
 		this.laborHrs = laborHrs;
 	}
