+++ src/main/java/com/agg/application/vo/ClaimPartVO.java
@@ -11,7 +11,7 @@ public class ClaimPartVO {
 	private int id;
 	private String partNo;
 	private String partDescr;
-	private int qty;
+	private double qty;
 	private double unitPrice;
 	/**
 	 * 
@@ -62,16 +62,19 @@ public class ClaimPartVO {
 	public void setPartDescr(String partDescr) {
 		this.partDescr = partDescr;
 	}
+
 	/**
 	 * @return the qty
 	 */
-	public int getQty() {
+	public double getQty() {
 		return qty;
 	}
+
+
 	/**
 	 * @param qty the qty to set
 	 */
-	public void setQty(int qty) {
+	public void setQty(double qty) {
 		this.qty = qty;
 	}
